backend/pyproject.toml
[tool.poetry]
name = "so-insights-back"
version = "0.1.0"
description = ""
authors = ["Samuel Mallet  <69467005+SuperMuel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beanie = "^1.26.0"
pymongo = "^4.8.0"
pydantic-settings = "^2.4.0"
streamlit = "^1.37.0"
uvicorn = "^0.30.5"
fastapi = {extras = ["standard"], version = "^0.112.0"}
arrow = "^1.3.0"


[tool.poetry.group.dev.dependencies]
make-it-sync = "^2.0.0"
pytest = "^8.3.2"
mongomock-motor = "^0.0.31"
setuptools = "^72.1.0"
ruff = "^0.5.6"
openapi-python-client = "^0.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
